+++ radish/customtyperegistry.py
@@ -1,0 +1,42 @@
+# -*- coding: utf-8 -*-
+
+"""
+    This module provides the Argument-Expression Registry
+"""
+
+from singleton import singleton
+
+from .exceptions import RadishError
+
+
+@singleton()
+class CustomTypeRegistry(object):
+    """
+        Registry for all custom argument expressions
+    """
+    def __init__(self):
+        self.custom_types = {}
+
+    def register(self, name, func):
+        """
+        Registers a custom type
+        """
+        if name in self.custom_types:
+            raise RadishError("Cannot register custom type with name {0} because it already exists".format(name))
+
+        self.custom_types[name] = func
+
+
+def custom_type(name, pattern):
+    """
+    Decorator for custom type pattern
+    """
+    def _decorator(func):
+        """
+        Actual decorator
+        """
+        func.pattern = pattern
+        CustomTypeRegistry().register(name, func)
+
+        return func
+    return _decorator
