+++ tests/functional/custom_types/radish/terrain.py
@@ -2,7 +2,7 @@
 
 from radish import before
 from radish import world
-from radish import arg_expr
+from radish import custom_type
 
 
 @before.each_scenario
@@ -10,8 +10,8 @@ def setup_hero_db(scenario):
     world.heros = []
 
 
-@arg_expr("Hero", r"[A-Z][a-z]+")
-def get_hero_arg_expr(text):
+@custom_type("Hero", r"[A-Z][a-z]+")
+def get_hero_custom_type(text):
     """
     Return a hero object by the given name
     """
